Refactor de las clases pojo & mix con las de domainModel

+++ src/tupleTypes/HabitacionsTuple.java
@@ -4,9 +4,9 @@ public class HabitacionsTuple {
 	
 	public String tipusHabitacio;
 	public int numeroDisponibles;
-	public float preu;
+	public double preu;
 
-	public HabitacionsTuple(String tipus, int numeroDisponibles, float preuTotal) {
+	public HabitacionsTuple(String tipus, int numeroDisponibles, double preuTotal) {
 		
 		this.tipusHabitacio = tipus;
 		this.numeroDisponibles = numeroDisponibles;
